// src/helpers/types.d.ts
export type GQLResponseType = {
    cursor: string | null;
    node: {
        id: string;
        tags: KeyValueType[];
        data: {
            size: string;
            type: string;
        };
    };
};
// ...
export type ArcGQLResponseType = {
    data: GQLResponseType[];
    nextCursor: string | null;
};
export type ArtifactResponseType = {
    nextCursor: string | null;
    previousCursor: string | null;
    contracts: GQLResponseType[];
};
export interface PoolType {
    id: string;
    state: PoolStateType;
}
export interface PoolStateType {
    title: string;
    image: string;
    briefDescription: string;
    description: string;
    link: string;
    owner: string;
    ownerInfo: string;
    timestamp: string;
    contributors: {
        [key: string]: string;
    };
    tokens: {
        [key: string]: string;
    };
    totalContributions: string;
    totalSupply: string;
    balance: string;
}
// ...
export type KeyValueType = {
    [key: string]: string | any;
};